fix Content-Length in bad compression cases
Occasionally a file ends up bigger than the original content when GZIP
compressed. However, the Content-Length in the HTTP header is reported
to be the original file size. This causes Chrome to truncate the content
when fetched via asynchronous AJAX. I ran into this issue when
developing in Emscripten and fetching texture files.

The correct fix is the choose the max of the original file length and
compressed content length.

I also refactored the compress method to be outside the class as I felt
that it doesn't have to actually belong in the class since it is purely
a util function.